DS-1359 Refactor Performance Tests to Reduce Code Complexity (#837)
# Task Branch Pull Request

**<https://nhsd-jira.digital.nhs.uk/browse/DS-1359>**

## Description of Changes

This PR changes the performance test from using separate classes for
each change event type to a single class
