+++ test/performance/stress_test.py
@@ -1,0 +1,5 @@
+from functions.performance_test import PerformanceTest
+
+
+class StressTest(PerformanceTest):
+    """This class is to run the stress test."""
