+++ test/pyproject.toml
@@ -28,10 +28,8 @@ target-version = "py311"
   "ARG001",
   "BLE001",
 ] # Conftest files break many rules.
-"performance/load_test_locustfile.py" = [
-  "ERA001",
-] # Allow commented out code in performance tests.
-"performance/stress_test_locustfile.py" = [
+
+"performance/functions/performance_test.py" = [
   "ERA001",
 ] # Allow commented out code in performance tests.
